fix: reconcile page JSON-LD with the homepage description fallback
main added a siteName fallback for empty homepage meta descriptions
(#3179), which getMetaDescription reads off props.site. This branch had
narrowed that helper to a site-less props type, so the merged tree failed
to compile. Include siteName in the site fields threaded into
getPageJsonLd instead.

Fold the JSON-LD tests into the __tests__/metadata.test.ts file that
#3179 introduced, so the module has a single test file again.
